fastapi/requirements.txt
fastapi
uvicorn
requests
redis
async_timeout
opentelemetry-distro
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-logging
